feat: 社員がスクラムチームに所属しているとき、社員を削除できないようにする

+++ packages/use-case/src/web/use-case/employee-use-case.ts
@@ -1,5 +1,11 @@
-import {Employee, EmployeeName, EmployeeRepositoryInterface, ID} from "@panda-project/core";
-import {EmployeeRepository} from "@/gateway";
+import {
+  Employee,
+  EmployeeName,
+  EmployeeRepositoryInterface,
+  ID,
+  ScrumTeamRepositoryInterface
+} from "@panda-project/core";
+import {EmployeeRepository, ScrumTeamRepository} from "@/gateway";
 import {AutoIncrementId} from "@/common";
 
 export class CreateEmployeeCommand {
@@ -45,6 +51,7 @@ export class DeleteEmployeeCommand {
 export class EmployeeUseCase {
   constructor(
     private readonly employeeRepository: EmployeeRepositoryInterface = new EmployeeRepository(),
+    private readonly scrumTeamRepository: ScrumTeamRepositoryInterface = new ScrumTeamRepository(),
   ) {
   }
 
@@ -65,6 +72,20 @@ export class EmployeeUseCase {
 
   async delete(command: DeleteEmployeeCommand) {
     const employee = await this.employeeRepository.findByIdOrFail(command.getEmployeeId())
+
+    let scrumTeam = null
+    try {
+      // TODO: fetch で null を返すようにする。null が返ってきて欲しいケースはここが初めてなので。
+      scrumTeam = await this.scrumTeamRepository.fetchOrFail()
+    } catch (e: unknown) {
+      await this.employeeRepository.delete(employee)
+      return
+    }
+
+    const isBelongToScrumTeam = scrumTeam.isBelongTo(employee.id)
+    if (isBelongToScrumTeam) {
+      throw new Error('社員がスクラムチームに所属しているため、削除できません')
+    }
 
     await this.employeeRepository.delete(employee)
   }
